Wrapped functions inside of Lambda
