+++ backend/hash_slinging_slasher/de_hasher/Cargo.toml
@@ -10,3 +10,10 @@ serde_json = "1.0.107"
 serde = { version = "1.0.188", features = ["derive"] }
 flate2 = "1.0.28"
 bincode="1.3.3"
+tokio = { version = "1", features = ["macros"] }
+tracing = { version = "0.1", features = ["log"] }
+tracing-subscriber = { version = "0.3", default-features = false, features = ["fmt"] }
+reqwest = {version = "0.11.22", features = ['json']}
+urlencoding = "2.1.3"
+lambda_http = "0.8.1"
+lambda_runtime = "0.8.1"
